src/gui/requirements.txt
PyQt5
pillow
scikit-image
numpy
h5py
tensorflow